# laravelProject/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Gestión de Vehículos",
        "description": "API RESTful para gestionar marcas de vehículos, personas y su relación de propiedad con vehículos. Incluye operaciones CRUD completas y endpoints de relaciones.",
        "contact": {
            "email": "dumar.pabon@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Servidor de la API"
        }
    ],
    "paths": {
...
        "/api/personas": {
            "get": {
                "tags": [
                    "Personas"
                ],
                "summary": "Lista todas las personas",
                "operationId": "19153d93fbbf375afea8d2a21f70831b",
                "responses": {
                    "200": {
                        "description": "Lista de personas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Persona"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Personas"
                ],
                "summary": "Crea una nueva persona",
                "operationId": "b76ce2b01c2e821292fac2c4cddb74ee",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PersonaCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Persona creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Persona"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/personas/{id}": {
            "get": {
                "tags": [
                    "Personas"
                ],
                "summary": "Obtiene una persona por ID",
                "operationId": "9ae93a77b50e534e16e3a4aded57564d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Persona encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Persona"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Persona no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Personas"
                ],
                "summary": "Actualiza una persona",
                "operationId": "a4370a9476fe98f54efa656fbddd395b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PersonaUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Persona actualizada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Persona"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Persona no encontrada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Personas"
                ],
                "summary": "Elimina una persona",
                "operationId": "65acb26d6a487dcc8cb74f199a297fe0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Persona eliminada exitosamente"
                    },
                    "404": {
                        "description": "Persona no encontrada"
                    }
                }
            }
        },
        "/api/personas/{id}/vehiculos": {
            "get": {
                "tags": [
                    "Personas"
                ],
                "summary": "Obtiene los vehículos de una persona",
                "operationId": "d6d8e49339a4a0393c50865a60b3f053",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de vehículos de la persona",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Vehiculo"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Persona no encontrada"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Persona": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Juan Pérez"
                    },
                    "cedula": {
                        "type": "string",
                        "example": "1234567890"
                    }
                },
                "type": "object"
            },
            "PersonaCreate": {
                "required": [
                    "nombre",
                    "cedula"
                ],
                "properties": {
                    "nombre": {
                        "type": "string",
                        "example": "Juan Pérez"
                    },
                    "cedula": {
                        "type": "string",
                        "example": "1234567890"
                    }
                },
                "type": "object"
            },
            "PersonaUpdate": {
                "properties": {
                    "nombre": {
                        "type": "string",
                        "example": "Juan Pérez"
                    },
                    "cedula": {
                        "type": "string",
                        "example": "1234567890"
                    }
                },
                "type": "object"
            },
            "Vehiculo": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "modelo": {
                        "type": "string",
                        "example": "Corolla"
                    },
                    "marca_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "numero_puertas": {
                        "type": "integer",
                        "example": 4
                    },
                    "color": {
                        "type": "string",
                        "example": "Rojo"
                    },
                    "propietarios_ids": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "example": [
                            1,
                            2
                        ]
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}